third_party/blink/web_tests/media/W3C/audio/events/event_order_loadedmetadata_loadeddata-expected.txt

setting src attribute on autoplay video should trigger loadedmetadata then loadeddata event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
